+++ azure-communication-ui/chat/src/main/java/com/azure/android/communication/ui/chat/redux/middleware/sdk/ChatActionHandler.kt
@@ -96,9 +96,8 @@ internal class ChatActionHandler(private val chatService: ChatService) {
                     ChatAction.MessageSent(
                         messageInfoModel = action.messageInfoModel,
                         id = result.id,
-                        )
+                    )
                 )
-
             }
         }
     }
