+++ azure-communication-ui/chat/src/main/java/com/azure/android/communication/ui/chat/repository/storage/MessageRepositorySkipListDelegate.kt
@@ -4,21 +4,21 @@
 package com.azure.android.communication.ui.chat.repository.storage
 
 import com.azure.android.communication.ui.chat.models.EMPTY_MESSAGE_INFO_MODEL
-import com.azure.android.communication.ui.chat.models.INVALID_INDEX
 import com.azure.android.communication.ui.chat.models.MessageInfoModel
 import com.azure.android.communication.ui.chat.repository.IMessageRepositoryDelegate
-import java.util.TreeMap
+import java.util.concurrent.ConcurrentSkipListMap
 
-internal class IMessageRepositoryTreeDelegate : IMessageRepositoryDelegate {
+internal class MessageRepositorySkipListDelegate : IMessageRepositoryDelegate {
 
-    private val treeMapStorage: TreeMap<Long, MessageInfoModel> = TreeMap()
+    private val skipListStorage: ConcurrentSkipListMap<Long, MessageInfoModel> =
+        ConcurrentSkipListMap()
 
     override val size: Int
-        get() = treeMapStorage.size
+        get() = skipListStorage.size
 
     override fun addMessage(messageInfoModel: MessageInfoModel) {
         val orderId: Long = messageInfoModel.normalizedID
-        treeMapStorage[orderId] = messageInfoModel
+        skipListStorage[orderId] = messageInfoModel
     }
 
     override fun addPage(page: List<MessageInfoModel>) {
@@ -28,22 +28,27 @@ internal class IMessageRepositoryTreeDelegate : IMessageRepositoryDelegate {
     override fun removeMessage(message: MessageInfoModel) {
         val orderId = message.normalizedID
 
-        if (treeMapStorage.contains(orderId)) {
-            treeMapStorage.remove(orderId)
+        if (skipListStorage.contains(orderId)) {
+            skipListStorage.remove(orderId)
         }
+    }
+
+    override fun replaceMessage(oldMessage: MessageInfoModel, newMessage: MessageInfoModel) {
+        removeMessage(oldMessage)
+        addMessage(newMessage)
     }
 
     private fun searchItem(kth: Int): MessageInfoModel {
 
-        var highestKey = treeMapStorage.lastKey()
-        var lowestKey = treeMapStorage.firstKey()
+        var highestKey = skipListStorage.lastKey()
+        var lowestKey = skipListStorage.firstKey()
         var elements = 0
         var midKey: Long = 0
         var items = kth
         while (lowestKey <= highestKey) {
             midKey = (highestKey + lowestKey).div(2)
 
-            elements = treeMapStorage.subMap(lowestKey, midKey + 1).size
+            elements = skipListStorage.subMap(lowestKey, midKey + 1).size
             if (elements < items) {
                 items -= elements
                 lowestKey = midKey + 1
@@ -54,13 +59,13 @@ internal class IMessageRepositoryTreeDelegate : IMessageRepositoryDelegate {
             }
         }
 
-        val key = treeMapStorage.subMap(lowestKey, midKey + 1).lastKey()
-        return treeMapStorage.get(key)!!
+        val key = skipListStorage.subMap(lowestKey, midKey + 1).lastKey()
+        return skipListStorage.get(key)!!
     }
 
     fun searchIndexByID(messageId: Long): Int {
-        var highestKey = treeMapStorage.lastKey()
-        var lowestKey = treeMapStorage.firstKey()
+        var highestKey = skipListStorage.lastKey()
+        var lowestKey = skipListStorage.firstKey()
         var midKey: Long = 0
 
         while (lowestKey <= highestKey) {
@@ -74,7 +79,7 @@ internal class IMessageRepositoryTreeDelegate : IMessageRepositoryDelegate {
                 break
             }
         }
-        return treeMapStorage.headMap(midKey).size
+        return skipListStorage.headMap(midKey).size
     }
 
     override fun get(index: Int): MessageInfoModel = try {
